# src/pluma/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Pluma Diagnostic API
  version: 0.1.0
  summary: Diagnose LLM-mediated system failures from eval and observability data.
  description: |
    Pluma reads failing rows from eval and observability platforms (Braintrust,
    LangSmith, PostHog) and surfaces them as structured findings against the
    [agent-diagnosis-spec v0.2](https://github.com/ivaylogb/agent-diagnosis-spec)
    `FailingEvalContainer` schema.

    Each finding carries a claim and a category, with optional `file:line`
    citation and an applyable edit. Findings are categorized into one of
    four failure modes: `measurement_instrument`, `interface`,
    `context_at_decision`, or `call_sequence`. See the v0.1 capability
    section below for which fields are populated in the current version.

    ## Resource model

    The API exposes a single primary resource: `Job`. Create a diagnostic job
    with `POST /v1/jobs`. Poll its state with `GET /v1/jobs/{job_id}`. Once
    `status` is `completed`, fetch the findings sub-resource via
    `GET /v1/jobs/{job_id}/findings`. Cancel in-flight jobs with
    `DELETE /v1/jobs/{job_id}`.

    Diagnostic runs are asynchronous because the underlying triage loop calls
    out to an LLM and may take several minutes.

    ## Auth

    All requests except `GET /v1/healthz` require an API key in the
    `X-Pluma-Key` header. The key is generated at server startup and printed
    once to stderr.

    ## Request IDs

...

...

...

...
    |---|---|---|
...

...

...

  contact:
...
servers:
...
security:
  - apiKey: []
...
paths:
  /v1/jobs:
    post:
      tags: [jobs]
      operationId: createJob
      summary: Create a new diagnostic job.
      description: |
        Accepts one or more eval/observability sources and queues a diagnostic
        job. Returns immediately with a `job_id` and `status: "pending"`.

        Poll `GET /v1/jobs/{job_id}` until status is `completed`, `failed`, or
        `cancelled`. Then fetch findings via `GET /v1/jobs/{job_id}/findings`.

        ## Idempotency

        Submitting the same `Idempotency-Key` within 24 hours returns the
        existing job with a `200 OK` response. First-time submissions return
        `202 Accepted`. The status code itself signals whether the response
        is a fresh creation or an idempotent replay.

        Without an idempotency key, every call creates a new job.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
            examples:
              braintrust_only:
                summary: Single Braintrust experiment
                value:
                  sources:
                    - type: braintrust
                      experiment_id: exp_abc123
                      primary_scorer: factuality
              multi_source:
                summary: Braintrust + LangSmith combined
                value:
                  sources:
                    - type: braintrust
                      experiment_id: exp_abc123
                      primary_scorer: factuality
                    - type: langsmith
                      project_name: my-agent-prod
                      filter: 'eq(feedback_key, "thumbs_down")'
                  agent_revision: git-sha-abc123
      responses:
        '200':
          description: |
            Idempotent replay. A job with this `Idempotency-Key` already
            exists; the response body is its current state.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '202':
          description: New job created and queued.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                status: pending
                created_at: '2026-05-18T12:00:00Z'
                terminated_at: null
                error: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/InternalError'
  /v1/jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobIdParam'
    get:
      tags: [jobs]
      operationId: retrieveJob
      summary: Retrieve a diagnostic job.
      description: |
        Returns the current state of a diagnostic job. Use this endpoint to
        poll until the job reaches a terminal status (`completed`, `failed`,
        or `cancelled`).

        Recommended polling interval is 2-5 seconds. The server does not
        currently enforce rate limits but agents and SDKs should not poll
        more aggressively than once per second.

        The `error` field is non-null only when `status == "failed"`. For all
        other statuses the field is null.
      responses:
        '200':
          description: Job state retrieved.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              examples:
                pending:
                  summary: Job not yet started
                  value:
                    job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                    status: pending
                    created_at: '2026-05-18T12:00:00Z'
                    terminated_at: null
                    error: null
                running:
                  summary: Job in progress
                  value:
                    job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                    status: running
                    created_at: '2026-05-18T12:00:00Z'
                    terminated_at: null
                    error: null
                completed:
                  summary: Job complete and findings available
                  value:
                    job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                    status: completed
                    created_at: '2026-05-18T12:00:00Z'
                    terminated_at: '2026-05-18T12:03:42Z'
                    error: null
                failed:
                  summary: Job failed during execution
                  value:
                    job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                    status: failed
                    created_at: '2026-05-18T12:00:00Z'
                    terminated_at: '2026-05-18T12:01:15Z'
                    error:
                      code: source_fetch_failed
                      message: Braintrust API returned 404 for experiment exp_abc123
                      request_id: req_B7T7N835B1AX6A795C0TWB1E
                      details:
                        source_type: braintrust
                        experiment_id: exp_abc123
                cancelled:
                  summary: Job cancelled before completion
                  value:
                    job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                    status: cancelled
                    created_at: '2026-05-18T12:00:00Z'
                    terminated_at: '2026-05-18T12:00:30Z'
                    error: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [jobs]
      operationId: cancelJob
      summary: Cancel a diagnostic job.
      description: |
        Cancels a job in `pending` or `running` state. Returns the job with
        status `cancelled`.

        Cancelling a job already in a terminal state (`completed`, `failed`,
        `cancelled`) is a no-op and returns the existing terminal state
        without raising an error. The semantics are idempotent: calling
        `DELETE` repeatedly is safe.
      responses:
        '200':
          description: Job cancelled, or already in a terminal state.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /v1/jobs/{job_id}/findings:
    parameters:
      - $ref: '#/components/parameters/JobIdParam'
    get:
      tags: [jobs]
      operationId: retrieveJobFindings
      summary: Retrieve findings for a completed diagnostic job.
      description: |
        Returns the structured findings for a completed job. The `findings`
        field is a `FailingEvalContainer` matching the [agent-diagnosis-spec
        v0.2](https://github.com/ivaylogb/agent-diagnosis-spec) schema.

        Returns `409 Conflict` if the job is not yet in `completed` state.
        Returns `410 Gone` if the job completed but findings have expired
        (default retention: 24 hours from completion).
      responses:
        '200':
          description: Findings retrieved. `status` is always `completed`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobFindings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job is not yet complete.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: job_not_complete
                message: Job is in status 'running'. Wait for status 'completed' before fetching findings.
                request_id: req_B7T7N835B1AX6A795C0TWB1E
                details:
                  job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
                  current_status: running
        '410':
          description: Findings have expired and are no longer retrievable.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Pluma-Key
      description: |
        API key generated at server startup. Printed once to stderr on first
        boot. Persisted in `/var/lib/pluma/key` inside the container.
  headers:
    RequestId:
      description: |
        Unique identifier for this request, ULID-formatted with a `req_`
        prefix. Echoed back in error response bodies as `request_id`.
        Include this value when reporting issues.
      schema:
        type: string
        pattern: '^req_[0-9A-HJKMNP-TV-Z]{24}$'
        example: req_B7T7N835B1AX6A795C0TWB1E
  parameters:
    JobIdParam:
      name: job_id
      in: path
      required: true
      description: Unique identifier for a diagnostic job.
      schema:
        $ref: '#/components/schemas/JobId'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional client-generated key for idempotent retries. If a request
        with the same key arrives within 24 hours, the existing job is
        returned with `200 OK`.

        Keys are limited to ASCII alphanumerics, hyphens, and underscores.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]+$'
        minLength: 1
        maxLength: 64
        example: client-retry-abc123
  schemas:
    JobId:
      type: string
      pattern: '^job_[0-9A-HJKMNP-TV-Z]{24}$'
      description: |
        Stable identifier for a diagnostic job. ULID-formatted with a `job_`
        prefix and Crockford base32 encoding (no ambiguous characters).
      example: job_01HPCZ5VB5AQ8H21W2T5EZYP
    CreateJobRequest:
      type: object
      required: [sources]
      properties:
        sources:
          type: array
          minItems: 1
          maxItems: 10
          description: |
            One or more eval/observability sources to diagnose. At least one
            source is required. Multiple sources are processed together as a
            single diagnostic run.
          items:
            $ref: '#/components/schemas/Source'
        agent_revision:
          type: string
          maxLength: 64
          description: |
            Optional git SHA, version tag, or other revision identifier for
            the agent being diagnosed. When provided, this value overrides
            any agent revision auto-resolved from source metadata
            (e.g., Braintrust experiment metadata, LangSmith run tags), and
            applies to all sources in the request. Useful when sources
            disagree about which agent revision they captured.
          example: git-sha-abc123
    Source:
      oneOf:
        - $ref: '#/components/schemas/BraintrustSource'
        - $ref: '#/components/schemas/LangSmithSource'
        - $ref: '#/components/schemas/PostHogSource'
      discriminator:
        propertyName: type
        mapping:
          braintrust: '#/components/schemas/BraintrustSource'
          langsmith: '#/components/schemas/LangSmithSource'
          posthog: '#/components/schemas/PostHogSource'
    BraintrustSource:
      type: object
      required: [type, experiment_id]
      properties:
        type:
          type: string
          enum: [braintrust]
        experiment_id:
          type: string
          maxLength: 128
          description: |
            Braintrust experiment ID. Format is determined by Braintrust;
            Pluma does not validate the shape beyond a length limit.
          example: exp_abc123
        primary_scorer:
          type: string
          maxLength: 64
          description: |
            Scorer name to use for identifying failing rows in this Braintrust
            experiment. Falls back to the experiment's default scorer if not
            set.
          example: factuality
    LangSmithSource:
      type: object
      required: [type, project_name]
      properties:
        type:
          type: string
          enum: [langsmith]
        project_name:
          type: string
          maxLength: 128
          description: |
            LangSmith project name. Format is determined by LangSmith;
            Pluma does not validate the shape beyond a length limit.
          example: my-agent-prod
        filter:
          type: string
          maxLength: 512
          description: |
            LangSmith filter DSL expression. See the LangSmith docs for full
            syntax: https://docs.smith.langchain.com/observability/how_to_guides/filter_traces_in_application

            Common patterns:

            - `eq(feedback_key, "thumbs_down")` — failures flagged by user feedback
            - `and(eq(status, "error"), gte(start_time, "2026-05-01"))` — errors in a time window
            - `has(tags, "production")` — only production-tagged runs
            - `lt(latency, 0.5)` — runs faster than 500ms
          example: 'eq(feedback_key, "thumbs_down")'
        workflow:
          type: string
          enum: [dataset_experiment, project_traced]
          default: project_traced
          description: |
            Which LangSmith data shape to read. `dataset_experiment` reads
            from a dataset-and-experiment pairing; `project_traced` reads
            from production traces in a project.
    PostHogSource:
      type: object
      required: [type, cohort_id]
      properties:
        type:
          type: string
          enum: [posthog]
        cohort_id:
          type: string
          maxLength: 128
          description: |
            PostHog cohort or insight ID. Format is determined by PostHog;
            Pluma does not validate the shape beyond a length limit.
          example: cohort_456
    Job:
      type: object
      required: [job_id, status, created_at, terminated_at, error]
      properties:
        job_id:
          $ref: '#/components/schemas/JobId'
        status:
          $ref: '#/components/schemas/JobStatus'
        created_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp when the job was submitted.
        terminated_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: |
            RFC 3339 timestamp when the job reached any terminal status
            (`completed`, `failed`, or `cancelled`). Null while the job is
            `pending` or `running`.
        error:
          oneOf:
            - $ref: '#/components/schemas/Error'
            - type: 'null'
          description: |
            Error details if the job failed. Null for all other statuses
            (`pending`, `running`, `completed`, `cancelled`).
    JobStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
      description: |
        Job lifecycle state:

        - `pending`: queued but not yet started
        - `running`: actively being processed
        - `completed`: terminal — findings available via the findings sub-resource
        - `failed`: terminal — see `error` field for details
        - `cancelled`: terminal — cancelled via `DELETE /v1/jobs/{job_id}`
    JobFindings:
      type: object
      required: [job_id, status, findings]
      properties:
        job_id:
          $ref: '#/components/schemas/JobId'
        status:
          $ref: '#/components/schemas/JobStatus'
        findings:
          $ref: '#/components/schemas/FailingEvalContainer'
    FailingEvalContainer:
      type: object
      required: [version, evals]
      description: |
        Root container for diagnostic findings, conforming to
        [agent-diagnosis-spec v0.2](https://github.com/ivaylogb/agent-diagnosis-spec).
      properties:
        version:
          type: string
          enum: ['0.2']
          description: Schema version. Always `0.2` for this API.
        evals:
          type: array
          description: One entry per diagnosed failing eval.
          items:
            $ref: '#/components/schemas/FailingEval'
        agent_revision:
          oneOf:
            - type: string
            - type: 'null'
          description: |
            Git SHA, version tag, or other revision identifier for the agent
            that produced the failing evals. Null if no revision could be
            resolved from sources or override.
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: |
            Stable error code. Safe to programmatically dispatch on. Known
            values: `bad_request`, `unauthorized`, `not_found`,
            `validation_failed`, `job_not_complete`, `source_fetch_failed`,
            `findings_expired`, `internal_error`.
          example: source_fetch_failed
        message:
          type: string
          description: Human-readable error message.
          example: Braintrust API returned 404 for experiment exp_abc123
        request_id:
          type: string
          pattern: '^req_[0-9A-HJKMNP-TV-Z]{24}$'
          description: |
            Unique identifier for the request that produced this error.
            Matches the `X-Request-Id` response header. Include this value
            when reporting issues.
          example: req_B7T7N835B1AX6A795C0TWB1E
        details:
          type: object
          additionalProperties: true
          description: |
            Optional structured context. Field names depend on the error code:

            - `validation_failed` includes `field` (the offending request path)
              and `error` (the validation rule that failed)
            - `source_fetch_failed` includes `source_type` and a
              source-specific identifier (`experiment_id`, `project_name`, etc.)
            - `not_found` includes the missing identifier (`job_id`, etc.)
            - `job_not_complete` includes `job_id` and `current_status`
  responses:
    BadRequest:
      description: Request was malformed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: bad_request
            message: Request body is not valid JSON
            request_id: req_B7T7N835B1AX6A795C0TWB1E
    Unauthorized:
      description: Missing or invalid API key.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: unauthorized
            message: Missing or invalid X-Pluma-Key header
            request_id: req_B7T7N835B1AX6A795C0TWB1E
    NotFound:
      description: Resource not found.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: not_found
            message: No job with id job_01HPCZ5VB5AQ8H21W2T5EZYP
            request_id: req_B7T7N835B1AX6A795C0TWB1E
            details:
              job_id: job_01HPCZ5VB5AQ8H21W2T5EZYP
    ValidationError:
      description: Request body failed schema validation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: validation_failed
            message: Request body failed schema validation
            request_id: req_B7T7N835B1AX6A795C0TWB1E
            details:
              field: sources[0].experiment_id
              error: required field missing
    InternalError:
      description: |
        An unexpected server error occurred. The `code` field will be
        `internal_error`. Retry with exponential backoff; if the error
        persists, report the `request_id`.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: An unexpected error occurred. Please report request_id req_B7T7N835B1AX6A795C0TWB1E.
            request_id: req_B7T7N835B1AX6A795C0TWB1E